+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 2.0.0
+ * STORM-1268: port builtin-metrics to java
+ * STORM-1648: drpc spout reconnect on failure
  * STORM-1631: Storm CGroup bugs
  * STORM-1616: Add RAS API for Trident
  * STORM-1623: nimbus.clj's minor bug
@@ -63,6 +65,7 @@
  * STORM-1521: When using Kerberos login from keytab with multiple bolts/executors ticket is not renewed in hbase bolt.
 
 ## 1.0.0
+ * STORM-1625: Move storm-sql dependencies out of lib folder
  * STORM-1556: nimbus.clj/wait-for-desired-code-replication wrong reset for current-replication-count-jar in local mode
  * STORM-1636: Supervisor shutdown with worker id pass in being nil
  * STORM-1602: Blobstore UTs are failed on Windows